requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
supabase==2.3.4
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]==2.6.1
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx>=0.24.0,<0.26.0
numpy==1.26.2
pandas==2.1.1
scikit-learn>=1.3.2
joblib>=1.3.2
ucimlrepo>=0.0.7
